(GA) atmo_evolution manages infinite version of phase screens and atmo_infinte_evolution removed
