+++ config/other/params_atmo_test.yml
@@ -52,12 +52,13 @@ atmoF:
   outputs: ['layer_list']
 
 atmoI:
-  class:                'AtmoInfiniteEvolution'
+  class:                'AtmoEvolution'
   simul_params_ref:  'main'
   L0:                   [8] #, 20, 30, 10]                   # [m] Outer scale
   heights:              [119.] #, 180, 300, 500] #,837,3045,12780]), # [m] layer heights at 0 zenith angle
   Cn2:                  [1] #, 0.3, 0.1, 0.2] #,0.06,0.14,0.10]), # Cn2 weights (total must be eq 1)
   fov:                  0.0
+  infinite_ps:          True
   inputs:
     seeing: 'seeing.output'
     wind_speed: 'wind_speed.output'
