+++ specula/scripts/parse_classes.py
@@ -6,7 +6,7 @@ import sys
 # All the classes are:
 # AtmoEvolution.yml          BaseOperation.yml      DataStore.yml      FlaskServer.yml    ImRecCalibrator.yml           
 # ModalAnalysisWFS.yml  MultiImRecCalibrator.yml    Slopec.yml
-# AtmoInfiniteEvolution.yml  BaseProcessingObj.yml  DisplayServer.yml  FuncGenerator.yml  Integrator.yml                
+# BaseProcessingObj.yml  DisplayServer.yml  FuncGenerator.yml  Integrator.yml                
 # ModalAnalysis.yml     ProcessingContainer.yml   ShSlopec.yml           SnCalibrator.yml
 # AtmoPropagation.yml        CCD.yml                DM.yml             IdealWFS.yml       LowPassFilter.yml             
 # Modalrec.yml          PSF.yml                   ShSubapCalibrator.yml  Vibrations.yml
@@ -16,7 +16,7 @@ import sys
 
 # TODO new generator classes not 
 exposed_classes = [ 'Source', 'Pupilstop',                    
-                    'FuncGenerator', 'BaseOperation', 'AtmoEvolution', 'AtmoInfiniteEvolution', 'AtmoPropagation',
+                    'FuncGenerator', 'BaseOperation', 'AtmoEvolution', 'AtmoPropagation',
                     'ModulatedPyramid', 'CCD', 'Slopec', 'PyrSlopec', 'Modalrec', 'Integrator', 'IirFilter', 'DM', 'PSF', 'DataStore'                
                   ]
